refactor: getTokens 경로를 sessions 디렉토리로 수정
