+++ src/entities/auth/services/types.ts
@@ -12,3 +12,9 @@ export interface LoginResponse {
 export interface SignupResponse {
     message: string;
 }
+
+export interface UserTokens {
+    idToken: string;
+    accessToken: string;
+    refreshToken: string;
+}
